refactor: remove index files
